test: refactor `InlinePhrasingContent/Text` test-case
