--- packages/repixe-parse/test/pxast.test.ts
@@ -1,113 +1,0 @@
-import type { Root } from "@rshirohara/pxast";
-import { describe, expect, test } from "vitest";
-import { fromPixivNovel } from "~/lib/index.js";
-
-describe("PhrasingStaticContent", () => {
-  describe("Text", () => {
-    test("ただのテキストはちゃんとただのテキストになる", () => {
-      const source = "ただのテキスト";
-      const expected: Root = {
-        type: "root",
-        children: [
-          {
-            type: "paragraph",
-            children: [{ type: "text", value: "ただのテキスト" }],
-          },
-        ],
-      };
-      expect(fromPixivNovel(source)).toEqual(expected);
-    });
-  });
-});
-
-describe("空要素の除去", () => {
-  describe("Paragraph", () => {
-    describe("Break", () => {
-      test("Paragraph の先頭にある Break は削除される", () => {
-        const source = "\n一行目\n二行目\n\n\n二段落目\n\n三段落目";
-        const expected: Root = {
-          type: "root",
-          children: [
-            {
-              type: "paragraph",
-              children: [
-                { type: "text", value: "一行目" },
-                { type: "break" },
-                { type: "text", value: "二行目" },
-              ],
-            },
-            {
-              type: "paragraph",
-              children: [{ type: "text", value: "二段落目" }],
-            },
-            {
-              type: "paragraph",
-              children: [{ type: "text", value: "三段落目" }],
-            },
-          ],
-        };
-        expect(fromPixivNovel(source)).toEqual(expected);
-      });
-      test("Paragraph の末尾にある Break は削除される", () => {
-        const source = "一行目\n二行目\n\n二段落目\n二行目\n\n\n三段落目\n";
-        const expected: Root = {
-          type: "root",
-          children: [
-            {
-              type: "paragraph",
-              children: [
-                { type: "text", value: "一行目" },
-                { type: "break" },
-                { type: "text", value: "二行目" },
-              ],
-            },
-            {
-              type: "paragraph",
-              children: [
-                { type: "text", value: "二段落目" },
-                { type: "break" },
-                { type: "text", value: "二行目" },
-              ],
-            },
-            {
-              type: "paragraph",
-              children: [{ type: "text", value: "三段落目" }],
-            },
-          ],
-        };
-        expect(fromPixivNovel(source)).toEqual(expected);
-      });
-    });
-  });
-
-  describe("Link", () => {
-    describe("Text", () => {
-      test("Link の末尾にある空の Text は削除される", () => {
-        const source =
-          "[[jumpuri:リンクの[[rb:中>なか]]でルビが[[rb: 使用> しよう]]できる> https://example.com]]";
-        const expected: Root = {
-          type: "root",
-          children: [
-            {
-              type: "paragraph",
-              children: [
-                {
-                  type: "link",
-                  url: "https://example.com",
-                  children: [
-                    { type: "text", value: "リンクの" },
-                    { type: "ruby", value: "中", ruby: "なか" },
-                    { type: "text", value: "でルビが" },
-                    { type: "ruby", value: "使用", ruby: "しよう" },
-                    { type: "text", value: "できる" },
-                  ],
-                },
-              ],
-            },
-          ],
-        };
-        expect(fromPixivNovel(source)).toEqual(expected);
-      });
-    });
-  });
-});
